Tentative fix for bug found in benchmarking
The bug involved the fact that an expansion could be performed on a node that had just been removed - this is obviously improper behaviour. My first fix for this was to atomically
mark the node before expansopn, as this would prevent its removal. However this meant that if the node had since become an arraynode, it was now marked as both. The actual solution
is to use the compare_and_mark function on AtomicMarkable to make sure we know the node is still a data node before marking. This seems to fix the problem as the benchmark now works.

Yay! :)